+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# You're in Grave Danger 2.0.8
+
+---
+
 # You're in Grave Danger 2.0.7
 
 ### Changes